requirement.txt
asn1crypto==0.24.0
boto3==1.9.137
botocore==1.12.137
cryptography==2.1.4
djangorestframework==3.9.2
docutils==0.14
enum34==1.1.6
futures==3.2.0
idna==2.6
ipaddress==1.0.17
jmespath==0.9.4
keyring==10.6.0
keyrings.alt==3.0
mysql-connector-python==2.1.6
MySQL-python==1.2.5
mysql-utilities==1.6.4
paramiko==2.0.0
pexpect==4.2.1
pyasn1==0.4.2
pycrypto==2.6.1
pygobject==3.26.1
pyodbc==4.0.17
pysqlite==2.7.0
python-dateutil==2.8.0
pyxdg==0.25
s3transfer==0.2.0
SecretStorage==2.3.1
six==1.12.0
urllib3==1.24.2
